DragAndDrop 컨텐츠를 별도의 컴포넌트로 분리
